feat(setup): package setup setting
